+++ iOSDeviceManager/Commands/ResignCommand.m
@@ -1,3 +1,4 @@
+
 #import "Simulator.h"
 #import "ResignCommand.h"
 #import "DeviceUtils.h"