+++ iOSDeviceManager/Commands/iOSReturnStatusCode.h
@@ -1,3 +1,7 @@
+
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 typedef NS_ENUM(NSUInteger, iOSReturnStatusCode) {
     iOSReturnStatusCodeEverythingOkay = 0,
     iOSReturnStatusCodeGenericFailure,