+++ iOSDeviceManager/Devices/PhysicalDevice.h
@@ -1,4 +1,7 @@
 
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 @import Foundation;
 #import "Device.h"
 #import <XCTestBootstrap/XCTestBootstrap.h>