+++ iOSDeviceManager/Devices/Simulator.m
@@ -650,6 +650,12 @@ static const FBSimulatorControl *_control;
 - (iOSReturnStatusCode)isInstalled:(NSString *)bundleID {
 
     NSError *error = nil;
+    if (![self boot]) {
+        ConsoleWriteErr(@"Cannot check for installed application"
+                        "%@ on Simulator %@ because the device could not "
+                        "be booted", bundleID, self.fbSimulator);
+        return iOSReturnStatusCodeInternalError;
+    }
     BOOL installed = [self isInstalled:bundleID withError:&error];
 
     if (installed) {
@@ -711,6 +717,30 @@ static const FBSimulatorControl *_control;
     }
 
     [ConsoleWriter write:dest];
+    return iOSReturnStatusCodeEverythingOkay;
+}
+
+- (iOSReturnStatusCode)downloadXCAppDataBundleForApplication:(NSString *)bundleID
+                                                      toPath:(NSString *)path {
+    NSError *e;
+    NSString *containerPath = [self containerPathForApplication:bundleID];
+    if (!containerPath) {
+        ConsoleWriteErr(@"Unable to find container path for app %@ on device %@",
+                        bundleID, [self uuid]);
+        return iOSReturnStatusCodeGenericFailure;
+    }
+    if (![[NSFileManager defaultManager] removeItemAtPath:path error:&e]) {
+        ConsoleWriteErr(@"Error: %@", e.localizedDescription);
+        return iOSReturnStatusCodeInternalError;
+    }
+    if (![[NSFileManager defaultManager] copyItemAtPath:containerPath
+                                                 toPath:path
+                                                  error:&e]) {
+        ConsoleWriteErr(@"Unable to copy xcappdata for app %@ on device %@\n"
+                        "Error: %@",
+                        bundleID, [self uuid], e.localizedDescription);
+        return iOSReturnStatusCodeGenericFailure;
+    }
     return iOSReturnStatusCodeEverythingOkay;
 }