+++ iOSDeviceManager/Resigning/Certificate.h
@@ -1,3 +1,7 @@
+
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 #import <Foundation/Foundation.h>
 
 // TODO: check to see if the certificate has expired or has been revoked