+++ iOSDeviceManager/Utilities/MachClock.m
@@ -1,3 +1,4 @@
+
 #include <mach/mach.h>
 #include <mach/mach_time.h>
 #include "MachClock.h"